+++ src/components/Order/Stepper.styles.js
@@ -1,0 +1,26 @@
+export const containerStyle = {
+  width: "100%",
+};
+
+export const stepContentBox = {
+  my: 4,
+};
+
+export const buttonContainer = {
+  display: "flex",
+  flexDirection: "row",
+  pt: 2,
+};
+
+export const spacer = {
+  flex: "1 1 auto",
+};
+
+export const backButton = {
+  mr: 1,
+};
+
+export const completeText = {
+  mt: 2,
+  mb: 1,
+};
